+++ framework-crates/objc2-gl-kit/src/lib.rs
@@ -1,0 +1,35 @@
+//! # Bindings to the `GLKit` framework
+//!
+//! See [Apple's docs][apple-doc] and [the general docs on framework crates][framework-crates] for more information.
+//!
+//! [apple-doc]: https://developer.apple.com/documentation/glkit/
+//! [framework-crates]: https://docs.rs/objc2/latest/objc2/topics/about_generated/index.html
+#![no_std]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+// Update in Cargo.toml as well.
+#![doc(html_root_url = "https://docs.rs/objc2-gl-kit/0.3.1")]
+
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
+#[cfg(feature = "std")]
+extern crate std;
+
+mod generated;
+#[allow(unused_imports, unreachable_pub)]
+pub use self::generated::*;
+
+// OpenGL/gltypes.h
+// Not re-exported by objc2-open-gl.
+#[allow(unused)]
+pub(crate) type GLboolean = u8;
+#[allow(unused)]
+pub(crate) type GLenum = u32;
+#[allow(unused)]
+pub(crate) type GLfloat = core::ffi::c_float;
+#[allow(unused)]
+pub(crate) type GLint = i32;
+#[allow(unused)]
+pub(crate) type GLsizei = i32;
+#[allow(unused)]
+pub(crate) type GLuint = u32;